Thêm response (BaseResponse, OrderResponse, ProductResponse), dịch vụ OrderDetailService, cập nhật controller, DTO, model và repository

+++ src/main/java/com/project/shopping_app/dtos/OrderDTO.java
@@ -29,7 +29,7 @@ public class OrderDTO {
   private String note;
 
   @JsonProperty("total_money")
-  private Float totalMoney;
+  private Double totalMoney;
 
   @JsonProperty("shipping_method")
   private String shippingMethod;
